Issue509 cbct speedup - failing tests due to deprecated python versions in the build

+++ core/kipl/UnitTests/tkiplbasetest/tst_kiplbase.cpp
@@ -258,7 +258,7 @@ void Tkiplbase::testHistogram()
     
     size_t cnt = 0UL;
 
-    cnt = std::accumulate(hist.begin(),hist.end(),0UL); 
+    cnt = std::accumulate(hist.begin(),hist.end(),static_cast<size_t>(0)); 
     
     QCOMPARE(cnt,img.Size());
 }
